Proper default Keybinding initialization
