+++ Editor/Program.cs
@@ -1,9 +1,9 @@
 #nullable enable
 using SharpDX.Direct3D11;
+using SilkWindows;
 using System.Diagnostics;
 using System.Globalization;
 using System.IO;
-using SilkWindows;
 using T3.Core.Compilation;
 using T3.Core.IO;
 using T3.Core.Resource;
@@ -12,6 +12,7 @@ using T3.Core.UserData;
 using T3.Editor.App;
 using T3.Editor.Compilation;
 using T3.Editor.Gui;
+using T3.Editor.Gui.Interaction;
 using T3.Editor.Gui.Interaction.Camera;
 using T3.Editor.Gui.Interaction.StartupCheck;
 using T3.Editor.Gui.Styling;
@@ -137,6 +138,7 @@ internal static class Program
         SharedResources.Initialize();
 
         Log.Debug("Initialize User Interface...");
+        KeyboardBinding.Initialize();
 
         bool forceRecompileProjects;
